Commit Process

+++ trainer.py
@@ -162,7 +162,7 @@ def trainer(end_epoch,epoch_num,model,dataloader,optimizer,device,ckpt,num_class
         t_masks = targets * overlap
         targets = targets.float()
 
-        loss_ce = ce_loss(x=outputs, y=targets)
+        loss_ce = ce_loss(x=outputs, y=targets.long())
         # loss_ce = ce_loss(outputs, targets[:].long())
         loss_dice = dice_loss(outputs, targets, softmax=True)
